splitting the share intent handling from the attachments helper - decouples the attachment callback
